Revert "use asset proxy url for image"
