+++ .env
@@ -8,4 +8,3 @@ SOLANA_KEYPAIR_PATH=./keypair.json
 NFT_STORAGE_API_ENDPOINT=https://api.nft.storage
 IPFS_ENDPOINT=https://nftstorage.link/ipfs
 CREDIT_SHEET=credits.toml
-ASSET_CDN = http://assets.holaplex.tools
