+++ api/Cargo.toml
@@ -49,7 +49,7 @@ package = "holaplex-hub-core"
 version = "0.1.0"
 git = "https://github.com/holaplex/hub-core"
 branch = "stable"
-features = ["kafka", "credits", "asset_proxy"]
+features = ["kafka", "credits"]
 
 [build-dependencies.hub-core-build]
 package = "holaplex-hub-core-build"
