+++ api/src/entities/collection_mints.rs
@@ -3,11 +3,8 @@
 use async_graphql::{ComplexObject, Context, Result, SimpleObject};
 use sea_orm::entity::prelude::*;
 
-use super::sea_orm_active_enums::CreationStatus;
-use crate::{
-    objects::{Collection, MetadataJson},
-    AppContext,
-};
+use super::{metadata_jsons, sea_orm_active_enums::CreationStatus};
+use crate::{objects::Collection, AppContext};
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
 #[sea_orm(table_name = "collection_mints")]
@@ -69,7 +66,7 @@ impl CollectionMint {
 
     /// The metadata json associated to the collection.
     /// [Metaplex v1.1.0 Standard](https://docs.metaplex.com/programs/token-metadata/token-standard)
-    async fn metadata_json(&self, ctx: &Context<'_>) -> Result<Option<MetadataJson>> {
+    async fn metadata_json(&self, ctx: &Context<'_>) -> Result<Option<metadata_jsons::Model>> {
         let AppContext {
             metadata_json_loader,
             ..
