+++ api/src/entities/metadata_jsons.rs
@@ -1,13 +1,17 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.11.0
 
-use async_graphql::Result;
+use async_graphql::{ComplexObject, Context, Result, SimpleObject};
 use sea_orm::entity::prelude::*;
+
+use super::metadata_json_attributes;
+use crate::AppContext;
 
 /// The collection's associated metadata JSON.
 /// ## References
 /// [Metaplex v1.1.0 Standard](https://docs.metaplex.com/programs/token-metadata/token-standard)
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, SimpleObject)]
 #[sea_orm(table_name = "metadata_jsons")]
+#[graphql(complex, concrete(name = "MetadataJson", params()))]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
     pub id: Uuid,
@@ -26,6 +30,22 @@ pub struct Model {
     pub animation_url: Option<String>,
     /// An optional URL where viewers can find more information on the NFT, such as the collection's homepage or Twitter page.
     pub external_url: Option<String>,
+}
+
+#[ComplexObject]
+impl Model {
+    // The NFT's associated attributes.
+    async fn attributes(
+        &self,
+        ctx: &Context<'_>,
+    ) -> Result<Option<Vec<metadata_json_attributes::Model>>> {
+        let AppContext {
+            metadata_json_attributes_loader,
+            ..
+        } = ctx.data::<AppContext>()?;
+
+        metadata_json_attributes_loader.load_one(self.id).await
+    }
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
