+++ api/src/lib.rs
@@ -29,7 +29,6 @@ use dataloaders::{
 use db::Connection;
 use hub_core::{
     anyhow::{Error, Result},
-    assets::AssetProxy,
     clap,
     consumer::RecvError,
     credits::CreditsClient,
@@ -205,7 +204,6 @@ pub struct AppState {
     pub credits: CreditsClient<Actions>,
     pub solana: Solana,
     pub nft_storage: NftStorageClient,
-    pub asset_proxy: AssetProxy,
 }
 
 impl AppState {
@@ -217,7 +215,6 @@ impl AppState {
         credits: CreditsClient<Actions>,
         solana: Solana,
         nft_storage: NftStorageClient,
-        asset_proxy: AssetProxy,
     ) -> Self {
         Self {
             schema,
@@ -226,7 +223,6 @@ impl AppState {
             credits,
             solana,
             nft_storage,
-            asset_proxy,
         }
     }
 }
