+++ api/src/objects/mod.rs
@@ -5,7 +5,7 @@ mod collection_creator;
 mod customer;
 mod drop;
 mod holder;
-mod metadata_json;
+mod metadata_json_input;
 mod project;
 mod wallet;
 
@@ -14,6 +14,6 @@ pub use collection_creator::CollectionCreator;
 pub use customer::Customer;
 pub use drop::Drop;
 pub use holder::Holder;
-pub use metadata_json::{MetadataJson, MetadataJsonInput};
+pub use metadata_json_input::MetadataJsonInput;
 pub use project::Project;
 pub use wallet::Wallet;
